gQL to TS ready to merge into dev
